v1.0.0 - без оплаты

+++ src/api/api.js
@@ -1,7 +1,6 @@
 import axios from "axios";
 import TokenService from "../services/token";
-
-const AUTH_API_URL = "https://localhost:5000/api/";
+import Constants from "../constants";
 
 const instance = axios.create({
   headers: {
@@ -17,42 +16,35 @@ instance.interceptors.request.use(
     }
     return config;
   },
-  (error) => {
-    return Promise.reject(error);
-  }
+  (error) => Promise.reject(error)
 );
 
 instance.interceptors.response.use(
-  (res) => {
-    return res;
-  },
+  (res) => res,
   async (err) => {
     const originalConfig = err.config;
-    if (err.response) {
-      if (
-        (err.response.status === 401 || err.response.data.code === "1") &&
-        !originalConfig._retry
-      ) {
-        originalConfig._retry = true;
+    if (
+      err.response &&
+      (err.response.status === 401 || err.response.data.code === "1") &&
+      !originalConfig._retry
+    ) {
+      originalConfig._retry = true;
 
-        try {
-          const rs = await instance.get(
-            AUTH_API_URL +
-              "authentication/refresh?token=" +
-              TokenService.getRefreshToken()
-          );
+      try {
+        const rs = await instance.get(
+          Constants.GATE_AWAY_API_URL +
+            "authentication/refresh?token=" +
+            TokenService.getRefreshToken()
+        );
 
-          if (rs.status === 401 || rs.data.code === "1") {
-            TokenService.removeUser();
-          } else if (rs.status === 200) {
-            await TokenService.setUser(rs.data.data);
-          }
-
-          return instance(originalConfig);
-        } catch (_error) {
+        if (rs.status === 401 || rs.data.code === "1")
           TokenService.removeUser();
-          return Promise.reject(_error);
-        }
+        else if (rs.status === 200) await TokenService.setUser(rs.data.data);
+
+        return instance(originalConfig);
+      } catch (error) {
+        TokenService.removeUser();
+        return Promise.reject(error);
       }
     }
